refactor(mcp): drain CRM sales-pipeline tools into a scoped pack
Move the customer-account / opportunity / quote tool domain
(list_customer_accounts, list_opportunities, get_opportunity, list_quotes,
create_customer_account, create_opportunity, create_quote) out of the
executeTool switch in apps/web/lib/mcp-tools.ts and into a scoped, lazily
dispatched pack registered via pack-registry.ts. Definitions, handlers, and
grant mirrors travel together; behavior is preserved (verbatim handler bodies,
same coercion/service delegation/envelopes/grants). Reads hit prisma with merge
tombstones excluded; writes delegate to @/lib/actions/crm.

Process-Spine-Decision: EP-8DC217EB plan §4 BET-4 — drain the crm-sales-pipeline tool domain into a scoped pack (registered via pack-registry.ts); ratchet baseline shrunk; behavior-preserving relocation.
