+++ scripts/mcp-tool-pack-baseline.json
@@ -27,16 +27,13 @@
   "contribute_to_hive",
   "create_backlog_item",
   "create_build_epic",
-  "create_customer_account",
   "create_digital_product",
   "create_epic",
   "create_licensing_readiness_issue",
   "create_marketing_asset_task",
   "create_marketing_automation_candidate",
   "create_marketing_campaign_brief",
-  "create_opportunity",
   "create_portal_pr",
-  "create_quote",
   "create_release_bundle",
   "create_scheduled_agent_task",
   "deploy_feature",
@@ -72,7 +69,6 @@
   "get_marketing_summary",
   "get_my_coworker_profile",
   "get_next_recommended_work",
-  "get_opportunity",
   "get_release_status",
   "get_thread_result",
   "grok_signin_start",
@@ -83,13 +79,10 @@
   "list_all_capability_needs",
   "list_backlog_items",
   "list_build_activity_since",
-  "list_customer_accounts",
   "list_epics",
   "list_my_capability_needs",
-  "list_opportunities",
   "list_patch_posture",
   "list_project_directory",
-  "list_quotes",
   "list_sandbox_files",
   "list_scheduled_agent_tasks",
   "list_wiki_overlay_drafts",
